--- a/main/src/ca/uwaterloo/flix/language/phase/jvm/GenFunctionAbstractClasses.scala
+++ b/main/src/ca/uwaterloo/flix/language/phase/jvm/GenFunctionAbstractClasses.scala
@@ -1,5 +1,6 @@
 /*
  * Copyright 2017 Magnus Madsen
+ * Copyright 2021 Jonathan Lindegaard Starup
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.