Part of #2591.
Fixes #2205.
Master can have between 5000 - 6000 methods at most (`@test def asdfasdf<i>(): Int = 42`)
This PR can have 15000 - 20000 of the above methods (so that would be around 2500 realistic one line methods compared to ~800 before)
# The hierachy
## Continuations
We have a hierchy of classes that represent functions. The topmost one is the continuation abstract class that exists for each erased return type of all functions.
```
public abstract class Cont$Int32 implements Runnable {
    public int result;

    public Cont$Int32() {
        super();
    }

    public abstract Cont$Int32 invoke();

    public final int unwind() {
        Cont$Int32 var1 = this;
        Cont$Int32 var2 = null;

        do {
            var2 = var1;
            var1 = var1.invoke();
        } while(var1 != null);

        return var2.result;
    }

    public final void run() {
        this.unwind();
        return;
    }
}
```
## Functions
The next layer is the function abstract classes that exists for each erased arrow type like `(Int, Object) -> Object`.
```
public abstract class Fn1$Obj$Int32 extends Cont$Int32 implements java.util.function.Function {
    public abstract Object arg0;

    public Fn1$Obj$Int32() {
        super();
    }
}

```
## Defs
The actual classes that represent each Flix named function like this one
```
def main(_args: Array[String]): Int & Impure =
    println("main");
    let r = {l = 12};
    r.l
```
```
public final class Def$main extends Fn1$Obj$Int32 {
    public Def$main() {
        super();
    }

    public final Cont.Int32 invoke() {
        Object var1 = this.arg0;
        113035 var10000 = new 113035();
        var10000.arg0 = "main";
        Unit var2 = (Unit)var10000.unwind();
        RecordExtend.Int32 var4 = new RecordExtend.Int32();
        var4.label = "l";
        var4.value = 12;
        var4.rest = RecordEmpty.INSTANCE;
        RecordExtend.Int32 var3 = var4;
        this.result = ((RecordExtend.Int32)(()var3).lookupField("l")).value;
        return null;
    }
}
```
## Closures
Closures represent annonymous functions of flix and are almost like defs except that they have additional arguments for their captured variables (this is println for unit)
```
public final class Clo$println$113035$120317 extends Fn1$Obj$Obj {
    public Object clo0;

    public final Cont.Obj invoke() {
        Object var1 = this.clo0;
        Object var2 = this.arg0;
        ((PrintStream)((PrintStream)var1)).println((String)((String)var2));
        this.result = Unit.INSTANCE;
        return null;
    }

    public Clo$println$113035$120317() {
        super();
    }
}
```
# Namespace class
The namespace classes were used primarily for supporting context, but are used in some other contexts. The current NS class just contains an erased version of each flix method like main. I think this can be removed but I will leave it for future PR.
```
public final class Ns {

    public static final int m_main(Object var0) {
        main var10000 = new main();
        var10000.arg0 = var0;
        return var10000.unwind();
    }

    public static final Object m_$bar$greater$114382(Object var0, Object var1) {
        114382 var10000 = new 114382();
        var10000.arg0 = var0;
        var10000.arg1 = var1;
        return var10000.unwind();
    }

    public static final Object m_println$113096$120878(Object var0) {
        120878 var10000 = new 120878();
        var10000.arg0 = var0;
        return var10000.unwind();
    }
}
```
# Calling protocol
When you need to call either a def or a closure, you need to:
1. put the object on the stack (pretty simple for defs, they have on constructor arguments) (closures will be done by constructing the object then assigning the `clo<i>` fields.
2. assign the `arg<i>` fields to their value
3. **(tailcall)** areturn the object or **(non-tailcall)** call unwind on the object (self-tail call ignored here)
This works because every snippet of code will at least be inside the unwind call of main. So in that loop, whenever you return a continuation, that will be called by the outer loop. If you do not return a continuation, then the outer unwind loop will take the value left on the stack and store it in the result and return null as the continuation. 

# Performance
I am not sure why this is slower than master, there should just be less code running 
## This PR
`JvmBackend                     988,5 ms (27,7%)`
`Generated 15.250.110 Bytes of code from 49312 lines of source code.`
`Throughput: min:  4.270, max: 13.152, avg: 10.665, median: 11.210`
## Master
`JvmBackend                     931,9 ms (27,0%)`
`Generated 30.521.313 Bytes of code from 49312 lines of source code.`
`Throughput: min:  3.894, max: 12.641, avg: 10.071, median: 10.731`
## Mixed
### Fixpoint
Looks like a x3 slowdown ish

Benchmark | Iterations(PR) | Time (avg)(PR) | Iterations(Master) | Time (avg)(Master)
:--- | ---: | ---: | ---: | ---:
Project (Int, Int) into A (n = 16) | 156641 | 12 us/op | 87481 | 39 us/op
Project (Int, Int) into A (n = 256) | 40298 | 351 us/op | 15529 | 739 us/op
Project (Int, Int) into A (n = 4096) | 1772 | 8106 us/op | 702 | 17972 us/op
Project (Int, Int) into A (n = 65536) | 83 | 178370 us/op | 33 | 461846 us/op
Project (Int, Int) into A (n = 1048576) | 3 | 3814208 us/op | 1 | 10572563 us/op
Select (Int, Int) from A (n = 16) | 3425 | 137 us/op | 2949 | 384 us/op
Select (Int, Int) from A (n = 256) | 12623 | 92 us/op | 6501 | 304 us/op
Select (Int, Int) from A (n = 4096) | 2348 | 340 us/op | 1313 | 2031 us/op
Select (Int, Int) from A (n = 65536) | 443 | 5634 us/op | 189 | 19602 us/op
Select (Int, Int) from A (n = 1048576) | 93 | 189332 us/op | 36 | 584463 us/op
Solve (Closure) (n = 64) | 2752 | 471 us/op | 1884 | 611 us/op
Solve (Closure) (n = 256) | 1225 | 536 us/op | 716 | 1248 us/op
Solve (Closure) (n = 1024) | 411 | 2081 us/op | 302 | 5145 us/op
Solve (Closure) (n = 4096) | 173 | 9655 us/op | 130 | 22315 us/op
Solve (Closure) (n = 16384) | 54 | 45462 us/op | 50 | 121387 us/op
Solve (Closure) (n = 65536) | 32 | 216349 us/op | 21 | 537175 us/op
Solve (Closure) (n = 262144) | 12 | 1039575 us/op | 5 | 2417424 us/op
Solve (Cross Product 2) (n = 256) | 6500 | 438 us/op | 4200 | 1261 us/op
Solve (Cross Product 2) (n = 1024) | 2117 | 2078 us/op | 1493 | 5453 us/op
Solve (Cross Product 2) (n = 4096) | 232 | 10061 us/op | 379 | 25230 us/op
Solve (Cross Product 2) (n = 16384) | 274 | 48014 us/op | 110 | 121280 us/op
Solve (Cross Product 2) (n = 65536) | 62 | 217007 us/op | 28 | 537784 us/op
Solve (Cross Product 2) (n = 262144) | 13 | 1030527 us/op | 6 | 2517479 us/op
Solve (Cross Product 2) (n = 1048576) | 3 | 4647183 us/op | 1 | 10322360 us/op
Solve (Cross Product 3) (n = 4096) | 1418 | 10018 us/op | 593 | 23726 us/op
Solve (Cross Product 3) (n = 32768) | 69 | 107745 us/op | 57 | 245260 us/op
Solve (Cross Product 3) (n = 262144) | 12 | 1059462 us/op | 5 | 2440081 us/op
Solve (Intersect) (n = 16) | 9456 | 53 us/op | 7627 | 139 us/op
Solve (Intersect) (n = 256) | 7417 | 519 us/op | 4851 | 1377 us/op
Solve (Intersect) (n = 4096) | 503 | 11823 us/op | 469 | 30216 us/op
Solve (Intersect) (n = 65536) | 50 | 236588 us/op | 23 | 596164 us/op
Solve (Intersect) (n = 1048576) | 2 | 4972549 us/op | 1 | 11904863 us/op
Compile Rules (n = 16) | 544 | 1051 us/op | 391 | 2871 us/op
Compile Rules (n = 256) | 54 | 16394 us/op | 41 | 45496 us/op
Compile Rules (n = 4096) | 9 | 267820 us/op | 5 | 766376 us/op
### List
Again maybe around x3 slowdown in avg 
Benchmark | Time(PR) | Time(Master)
:--- | ---: | ---:
BenchmarkList.benchmark01 | 18.348 ns. | 59.703 ns.
BenchmarkList.benchmark02 | 28.518 ns. | 99.222 ns.
BenchmarkList.benchmark03 | 19.812 ns. | 69.230 ns.
BenchmarkList.benchmark04 | 32.135 ns. | 73.338 ns.
BenchmarkList.benchmark05 | 36.359 ns. | 81.381 ns.
BenchmarkList.benchmark06 | 63.213 ns. | 204.847 ns.
BenchmarkList.benchmark07 | 43.340 ns. | 120.463 ns.
BenchmarkList.benchmark08 | 20.007 ns. | 60.135 ns.
BenchmarkList.benchmark09 | 28.185 ns. | 77.632 ns.
BenchmarkList.benchmark10 | 37.609 ns. | 143.977 ns.
BenchmarkList.benchmark11 | 59.794 ns. | 186.396 ns.
BenchmarkList.benchmark12 | 78.161 ns. | 274.330 ns.
BenchmarkList.benchmark13 | 28.706 ns. | 55.843 ns.
BenchmarkList.benchmark14 | 22.506 ns. | 47.742 ns.
BenchmarkList.benchmark15 | 55.695 ns. | 138.481 ns.
BenchmarkList.benchmark16 | 93.088 ns. | 295.441 ns.
BenchmarkList.benchmark17 | 22.742 ns. | 59.895 ns.
BenchmarkList.benchmark18 | 44.192 ns. | 87.396 ns.