@magnus-madsen 2021-11-25
What about the namespace classes?