@magnus-madsen 2021-11-25
Can you run `BenchmarkLIst` and some of the `BenchmarkDatalog`?