@magnus-madsen 2021-11-25
I am very excited about this! This means that we can finally have a proper concurrency story, its reduces bytecode size, and it significantly simplifies the generated code! Nice!