@magnus-madsen 2021-11-25
I see more substantial speedups:

```
New:   min:  5.887, max: 19.017, avg: 15.316, median: 16.796
Old:   min:  5.665, max: 17.232, avg: 14.797, median: 16.279
```