@magnus-madsen 2021-11-25
On a long run:
```
NEW:  JvmBackend                     799,5 ms (33,5%)
OLD:  JvmBackend                     934,7 ms (37,2%)
```

17% speedup