@magnus-madsen 2021-11-25
Question (but not intended to be fixed in this PR): It seems that sometimes we allocate a new Def object just to call unwind. But is this always necessary? Could we not call the method directly?