@JonathanStarup 2021-11-26
hmmm good point. it cant be done atm since unwind uses the instanced invoke. but we could make a version that doesnt use the instance. This would have to be in the function abstract class however. we could basically use normal jvm function calls for non-tail calls and only use objects where needed. that sounds pretty efficient.