@JonathanStarup 2021-11-26
> Do you feel comfortable with the current PR? Shall I merge it? You promise it won't break master ? ;)

i will have another check once i get to my desk later, ill ping you afterwards :)