@magnus-madsen 2021-11-26
Nice! Now lets watch https://arewefast.flix.dev :)